feet: image uploading with cloudinary

+++ backend/controllers/product.js
@@ -1,6 +1,14 @@
 const { validationResult } = require("express-validator");
 
-const Product = require("../models/Product")
+const Product = require("../models/Product");
+const {v2: cloudinary} = require('cloudinary');
+require("dotenv").config();
+
+cloudinary.config({
+    cloud_name: 'dyh5lh3p9',
+    api_key: '376528417874836',
+    api_secret: process.env.CLOUD_API,
+})
 
 
 exports.addNewProduct = async(req,res) =>{
@@ -81,8 +89,6 @@ exports.updateProduct = async(req,res) =>{
             message: errors.array()[0].msg,
         })
     };// validation fail or not
-   
-   
     try{
         const {
             product_name,
@@ -93,8 +99,6 @@ exports.updateProduct = async(req,res) =>{
             product_details,
             seller_id,
             product_id} = req.body;
-
-           
        if(req.userId.toString() !== seller_id){
             throw new Error("Authorization Failed!")
        }
@@ -122,16 +126,35 @@ exports.updateProduct = async(req,res) =>{
 
 exports.deleteProduct = async(req,res) =>{
     const {id} = req.params;
-    
-    
     try{
         const productDoc = await Product.findOne({_id: id});
-        
+        if(!productDoc){
+            return res.status(404).json({
+                isSuccess: true,
+                message: "Product not found!",
+               })
+        }
         if(req.userId.toString() !== productDoc.seller.toString()){
             throw new Error("Authorization Failed!")
-       }
+       };
+        if(productDoc.images && Array.isArray(productDoc.images)){
+            const deletePromise = productDoc.images.map((img)=> {
+                const publicId = img.substring(
+                    img.lastIndexOf("/") + 1,
+                    img.lastIndexOf(".")
+                )
+                return new Promise((resolve,reject) => {
+                    cloudinary.uploader.destroy(publicId,(err,result)=>{
+                       if(err){
+                           reject(new Error("Destroy failed!"));
+                       }else{
+                           resolve(result);
+                       }
+                    })});
+             });
+                await Promise.all(deletePromise)
+            }
          await Product.findByIdAndDelete(id);
-
          return res.status(202).json({
             isSuccess: true,
             message: "Product deleted!",
@@ -141,6 +164,82 @@ exports.deleteProduct = async(req,res) =>{
             isSuccess: false,
             message: err.message,
             }
-        )}
+        )}};
+
+exports.uploadProductImages = async(req,res) =>{
+    const productImages = req.files;
+    const productId = req.body.product_id;
+    let secureUrlArray = [];
+    try{
+        productImages.forEach((img) => {
+            cloudinary.uploader.upload(
+                img.path,async(err,result)=>{
+                if(!err){
+                    const url = result.secure_url;
+                    secureUrlArray.push(url);
+                    if(productImages.length === secureUrlArray.length){
+                     await Product.findByIdAndUpdate(productId,{$push : {images: secureUrlArray}})
+                     return res.status(200).json({
+                        isSuccess: true,
+                        message: "Product images uploaded",
+                        secureUrlArray,
+                       })
+                    };
+                }else{
+                    throw new Error("Cloud upload Failed!")}
+               }
+            )
+        });
+    }catch(err){
+        return res.status(404).json({
+            isSuccess: false,
+            message: err.message,
+            })
+    }
 }
 
+exports.getSavedImages = async(req,res) =>{
+    const {id} = req.params;
+    try{
+        const productDoc = await Product.findById(id).select("images");
+        if(!productDoc){
+            throw new Error("Product not found")
+        }
+        return res.status(200).json({
+            isSuccess: true,
+            message: "Product images saved.",
+            data : productDoc,
+        })
+        
+    }catch(err){
+        return res.status(404).json({
+            isSuccess: false,
+            message: err.message,
+            })
+    }
+}
+
+exports.deleteProductImages = async(req,res) =>{
+   
+   try{
+    const productId = req.params.productId;
+    const decodeImgToDelete = decodeURIComponent(req.params.imgToDelete);
+    await Product.findByIdAndUpdate(productId,{$pull: {images: decodeImgToDelete}})
+    const publicId = decodeImgToDelete.substring(
+        decodeImgToDelete.lastIndexOf("/") + 1,
+        decodeImgToDelete.lastIndexOf(".")
+    )
+    await cloudinary.uploader.destroy(publicId);
+    return res.status(200).json({
+        isSuccess: true,
+        message: "Product image destroyed.",
+       
+    })
+
+   }catch(err){
+    return res.status(404).json({
+        isSuccess: false,
+        message: err.message,
+        })
+   }
+}
